fix: 통합로직 배치사이즈 축소, db 어긋남 해소

+++ src/main/java/com/cast/caspedia/boardgame/service/BoardgameService.java
@@ -291,8 +291,8 @@ public class BoardgameService {
                 maxPlayers,
                 minPlayTime,
                 maxPlayTime,
-                minGeekWeight,
-                maxGeekWeight,
+                (float) minGeekWeight,
+                (float) maxGeekWeight,
                 pageable
         );
